refactor(ui_bridge): split monolithic UIBridge into modular files; preserve stable entrypoint and window.UIBridge
