+++ ROguelike 3_10_2025 filut jarjestyksessa luultavasti vikaa/core/bridge/ui_bridge/follower.js
@@ -1,0 +1,15 @@
+import { hasUI } from "./shared.js";
+
+export function isFollowerOpen() {
+  try { return !!(hasUI() && window.UI.isFollowerOpen && window.UI.isFollowerOpen()); } catch (_) { return false; }
+}
+
+export function showFollower(ctx, view) {
+  if (!hasUI()) return;
+  try { window.UI.showFollower && window.UI.showFollower(ctx, view); } catch (_) {}
+}
+
+export function hideFollower(ctx) {
+  if (!hasUI()) return;
+  try { window.UI.hideFollower && window.UI.hideFollower(); } catch (_) {}
+}
