+++ ROguelike 3_10_2025 filut jarjestyksessa luultavasti vikaa/core/bridge/ui_bridge/loot.js
@@ -1,0 +1,15 @@
+import { hasUI } from "./shared.js";
+
+export function showLoot(ctx, list) {
+  if (!hasUI()) return;
+  try { window.UI.showLoot && window.UI.showLoot(list || []); } catch (_) {}
+}
+
+export function hideLoot(ctx) {
+  if (!hasUI()) return;
+  try { window.UI.hideLoot && window.UI.hideLoot(); } catch (_) {}
+}
+
+export function isLootOpen() {
+  try { return !!(hasUI() && window.UI.isLootOpen && window.UI.isLootOpen()); } catch (_) { return false; }
+}
