+++ ROguelike 3_10_2025 filut jarjestyksessa luultavasti vikaa/core/bridge/ui_bridge/shared.js
@@ -1,0 +1,3 @@
+export function hasUI() {
+  return (typeof window !== "undefined" && window.UI);
+}
