+++ ROguelike 3_10_2025 filut jarjestyksessa luultavasti vikaa/docs/index.html
@@ -123,6 +123,7 @@
         { title: 'Engineering: GMBridge Modularization (Slice 1)', path: 'docs/gm_bridge_modularization.md' },
         { title: 'Engineering: Next Steps', path: 'docs/next_steps.md' },
         { title: 'Engineering: UIOrchestration Modularization (Slice A)', path: 'docs/ui_orchestration_modularization.md' },
+        { title: 'Engineering: UIBridge Modularization (Slice B)', path: 'docs/ui_bridge_modularization.md' },
 
         { title: 'Data Docs: Index', path: 'data/docs/README.md' },
         { title: 'Data Docs: Palette Schema', path: 'data/docs/palette_schema.md' },
@@ -163,6 +164,7 @@
         'docs/gm_bridge_modularization.md': 'Concrete file-split plan for core/bridge/gm_bridge.js (Slice 1).',
         'docs/next_steps.md': 'What to do next: QA gates, invariants, and the next modularization slice to pick.',
         'docs/ui_orchestration_modularization.md': 'UIOrchestration file split: new module layout + invariants.',
+        'docs/ui_bridge_modularization.md': 'UIBridge file split: new module layout + invariants.',
 
         'data/docs/README.md': 'Index of data documentation and palette guides.',
         'data/docs/palette_schema.md': 'Schema for palette JSON: keys, sections, and constraints.',
