--- ROguelike 3_10_2025 filut jarjestyksessa luultavasti vikoa/.gitignore
@@ -1,2 +1,0 @@
-# Ignore accidental typo folder created by editor/tooling
-/core/
